requirements.txt
streamlit
clickhouse-connect
pandas
numpy
json-tricks
python-dotenv
openai
